Jump/Crouch/NewPlayerChar
So, I couldn't figure out why the jump animation wasn't working, it worked for the blueprint original but not the duplicate I was using. I ended up just using the original from the animation pack because it had a functioning jump animation. Got the crouch to work too by adding an input mapping in the project editor.  Also set up the blueprint, now called CD_PlayerChar_V3, to shrink the collision capsule when the player crouches so that the character can fit places that they can't while standing.  Will need to set something up that prevents standing while in specific areas, standing up while in a tunnel or too short area the characters feet push through the floor.