Fixed crouching - mostly
Fixed crouching so that the player characters feet don't push through the floor if crouch button is released under a short structure. However, for now, if player releases crouch button while player character is in area to short to stand, the player remains in crouch position until player character leaves area and crouch is pressed again.