Mostly Fixed Crouch/added coffee item
Smoothed out crouch a bunch but there is still an issue when exiting tunnel after releasing crouch button early, the player will rise up but the next crouch is not smooth, all after that are though. Also added the coffee item and but a chest into the cabin in tutorial level. In addition to that, set up the cd_firstpersonchar to spawn with different energy levels depending on whether the player is spawning in the tutorial level or not.