Finished Store
Finished the functionality of the store. Still needs to be made prettier like the  player/container inventories.  Might add a wiggle room feature to store later, with wiggle room feature player can buy things that are more expensive then the credits they have as long as the difference is less than 10% of total amount of credits player has